Implement GetBlockRange network API
GetBlockRange is an asynchronous streaming API. Other APIs are typically
returned to the client as a single response to the request. This API
differs in that multiple responses can be streamed to the client.

Any updates for this API are returned as a GetBlock response. The stream
is signaled as finished to the client when the GetBlockRange response is
sent to the client.