Use receipts in the block
tx_merkle_root has been changed to receipt root. Blocks now store
receipts instead of transactions. A receipt represents an executed
transaction with the corresponding log that was produced.

Currently, nothing is using the log, but it is stored alongside the
block body. Transfer transactions will be the first to start using
the log as a test.